+++ docs/changelog.rst
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+- Add :meth:`gidgethub.abc.GitHubAPI.manage_rate_limit` as an overridable
+  hook, along with the :attr:`gidgethub.abc.GitHubAPI.requests_in_flight`
+  attribute, to let users implement their own rate-limit/backpressure
+  strategies
+  (`Issue #19 <https://github.com/gidgethub/gidgethub/issues/19>`_)
+
 - Add :attr:`gidgethub.HTTPException.headers` attribute
   (`PR #227 <https://github.com/gidgethub/gidgethub/pull/227>`_)